// src/app/shared/types/nav-paths.model.ts
export enum NavPathTypes {
  HOME = '',
  MEMBERS = 'members',
  MEMBERS_ADD = 'members/add',
  MEMBERS_EDIT = 'members/edit',
  SCHEDULE = 'schedule',
  NEWS = 'news',
  NEWS_COMPOSE = 'news/compose',
  NEWS_EDIT = 'news/edit',
  LONDON_CHESS_CHAMPION = 'london-chess-champion',
  PHOTO_GALLERY = 'photo-gallery',
  ABOUT = 'about',
  LOGIN = 'login',
  SIGN_UP = 'signup',
}